Coalesce StatementContext and ConcreteStatementContext into a single, final class.
