+++ core/src/main/java/org/jdbi/v3/StatementContext.java
@@ -17,8 +17,12 @@ import java.lang.reflect.Method;
 import java.lang.reflect.Type;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.LinkedHashSet;
 import java.util.Map;
 import java.util.Optional;
+import java.util.Set;
 import java.util.stream.Collector;
 
 import org.jdbi.v3.tweak.Argument;
@@ -30,8 +34,34 @@ import org.jdbi.v3.tweak.ColumnMapper;
  * to all statement customizers. This makes it possible to parameterize the processing of
  * the tweakable parts of the statement processing cycle.
  */
-public interface StatementContext
+public final class StatementContext
 {
+    private final JdbiConfig config;
+
+    private final Set<Cleanable> cleanables = new LinkedHashSet<>();
+
+    private String            rawSql;
+    private String            rewrittenSql;
+    private String            locatedSql;
+    private PreparedStatement statement;
+    private Connection        connection;
+    private Binding           binding;
+    private Class<?>          sqlObjectType;
+    private Method            sqlObjectMethod;
+    private boolean           returningGeneratedKeys;
+    private boolean           concurrentUpdatable;
+    private String[]          generatedKeysColumnNames;
+
+    /* visible for testing */
+    StatementContext() {
+        this(new JdbiConfig());
+    }
+
+    StatementContext(JdbiConfig config)
+    {
+        this.config = config;
+    }
+
     /**
      * Specify an attribute on the statement context
      *
@@ -40,7 +70,10 @@ public interface StatementContext
      *
      * @return previous value of this attribute
      */
-    Object setAttribute(String key, Object value);
+    public Object setAttribute(String key, Object value)
+    {
+        return config.statementAttributes.put(key, value);
+    }
 
     /**
      * Obtain the value of an attribute
@@ -49,7 +82,10 @@ public interface StatementContext
      *
      * @return the value of the attribute
      */
-    Object getAttribute(String key);
+    public Object getAttribute(String key)
+    {
+        return config.statementAttributes.get(key);
+    }
 
     /**
      * Obtain all the attributes associated with this context as a map. Changes to the map
@@ -57,7 +93,10 @@ public interface StatementContext
      *
      * @return a map f attributes
      */
-    Map<String, Object> getAttributes();
+    public Map<String, Object> getAttributes()
+    {
+        return config.statementAttributes;
+    }
 
     /**
      * Obtain a column mapper for the given type in this context.
@@ -65,7 +104,10 @@ public interface StatementContext
      * @param type the target type to map to
      * @return a ColumnMapper for the given type, or null if no column mapper is registered for the given type.
      */
-    Optional<ColumnMapper<?>> findColumnMapperFor(Type type);
+    public Optional<ColumnMapper<?>> findColumnMapperFor(Type type)
+    {
+        return config.mappingRegistry.findColumnMapperFor(type, this);
+    }
 
     /**
      * Obtain an argument for given value in this context
@@ -73,27 +115,46 @@ public interface StatementContext
      * @param value the argument value.
      * @return an Argument for the given value.
      */
-    Optional<Argument> findArgumentFor(Type type, Object value);
+    public Optional<Argument> findArgumentFor(Type type, Object value) {
+        return config.argumentRegistry.findArgumentFor(type, value, this);
+    }
 
     /**
      * Obtain a collector for the given type in this context
      * @param type the result type of the collector
      * @return a Collector for the given result type, or null if no collector factory is registered for this type.
      */
-    Optional<Collector<?, ?, ?>> findCollectorFor(Type type);
+    public Optional<Collector<?, ?, ?>> findCollectorFor(Type type) {
+        return config.collectorRegistry.findCollectorFor(type);
+    }
 
     /**
      * Returns the element type for the given container type, if available.
      * @param containerType the container type.
      */
-    Optional<Type> elementTypeFor(Type containerType);
+    public Optional<Type> elementTypeFor(Type containerType) {
+        return config.collectorRegistry.elementTypeFor(containerType);
+    }
+
+    void setRawSql(String rawSql)
+    {
+        this.rawSql = rawSql;
+    }
 
     /**
      * Obtain the initial sql for the statement used to create the statement
      *
      * @return the initial sql
      */
-    String getRawSql();
+    public String getRawSql()
+    {
+        return rawSql;
+    }
+
+    void setRewrittenSql(String rewrittenSql)
+    {
+        this.rewrittenSql = rewrittenSql;
+    }
 
     /**
      * Obtain the located and rewritten sql
@@ -102,7 +163,15 @@ public interface StatementContext
      *
      * @return the sql as it will be executed against the database
      */
-    String getRewrittenSql();
+    public String getRewrittenSql()
+    {
+        return rewrittenSql;
+    }
+
+    void setLocatedSql(String locatedSql)
+    {
+        this.locatedSql = locatedSql;
+    }
 
     /**
      * Obtain the located sql
@@ -111,7 +180,15 @@ public interface StatementContext
      *
      * @return the sql which will be passed to the statement rewriter
      */
-    String getLocatedSql();
+    public String getLocatedSql()
+    {
+        return locatedSql;
+    }
+
+    void setStatement(PreparedStatement stmt)
+    {
+        statement = stmt;
+    }
 
     /**
      * Obtain the actual prepared statement being used.
@@ -120,29 +197,92 @@ public interface StatementContext
      *
      * @return Obtain the actual prepared statement being used.
      */
-    PreparedStatement getStatement();
+    public PreparedStatement getStatement()
+    {
+        return statement;
+    }
+
+    void setConnection(Connection connection)
+    {
+        this.connection = connection;
+    }
 
     /**
      * Obtain the JDBC connection being used for this statement
      *
      * @return the JDBC connection
      */
-    Connection getConnection();
+    public Connection getConnection()
+    {
+        return connection;
+    }
 
-    Binding getBinding();
+    void setBinding(Binding b)
+    {
+        this.binding = b;
+    }
 
-    Class<?> getSqlObjectType();
+    public Binding getBinding()
+    {
+        return binding;
+    }
 
-    Method getSqlObjectMethod();
+    // TODO make default access or remove altogether
+    public void setSqlObjectType(Class<?> sqlObjectType)
+    {
+        this.sqlObjectType = sqlObjectType;
+    }
+
+    public Class<?> getSqlObjectType()
+    {
+        return sqlObjectType;
+    }
+
+    // TODO make default access or remove altogether
+    public void setSqlObjectMethod(Method sqlObjectMethod)
+    {
+        this.sqlObjectMethod = sqlObjectMethod;
+    }
+
+    public Method getSqlObjectMethod()
+    {
+        return sqlObjectMethod;
+    }
+
+    void setReturningGeneratedKeys(boolean b)
+    {
+        if (isConcurrentUpdatable() && b) {
+            throw new IllegalArgumentException("Cannot create a result set that is concurrent "
+                    + "updatable and is returning generated keys.");
+        }
+        this.returningGeneratedKeys = b;
+    }
 
     /**
      * Is the statement being generated expected to return the generated keys?
      */
-    boolean isReturningGeneratedKeys();
+    public boolean isReturningGeneratedKeys()
+    {
+        return returningGeneratedKeys || generatedKeysColumnNames != null && generatedKeysColumnNames.length > 0;
+    }
 
-    String[] getGeneratedKeysColumnNames();
+    public String[] getGeneratedKeysColumnNames()
+    {
+        if (generatedKeysColumnNames == null) {
+            return new String[0];
+        }
+        return Arrays.copyOf(generatedKeysColumnNames, generatedKeysColumnNames.length);
+    }
 
-    void addCleanable(Cleanable cleanable);
+    void setGeneratedKeysColumnNames(String[] generatedKeysColumnNames)
+    {
+        this.generatedKeysColumnNames = Arrays.copyOf(generatedKeysColumnNames, generatedKeysColumnNames.length);
+    }
+
+    public void addCleanable(Cleanable cleanable)
+    {
+        this.cleanables.add(cleanable);
+    }
 
     /**
      * Return if the statement should be concurrent updatable.
@@ -153,6 +293,29 @@ public interface StatementContext
      *
      * @return if the statement generated should be concurrent updatable.
      */
-    boolean isConcurrentUpdatable();
+    public boolean isConcurrentUpdatable() {
+        return concurrentUpdatable;
+    }
 
+    /**
+     * Set the context to create a concurrent updatable result set.
+     *
+     * This cannot be combined with {@link #isReturningGeneratedKeys()}, only
+     * one option may be selected. It does not make sense to combine these either, as one
+     * applies to queries, and the other applies to updates.
+     *
+     * @param concurrentUpdatable if the result set should be concurrent updatable.
+     */
+    void setConcurrentUpdatable(final boolean concurrentUpdatable) {
+        if (concurrentUpdatable && isReturningGeneratedKeys()) {
+            throw new IllegalArgumentException("Cannot create a result set that is concurrent "
+                    + "updatable and is returning generated keys.");
+        }
+        this.concurrentUpdatable = concurrentUpdatable;
+    }
+
+    public Collection<Cleanable> getCleanables()
+    {
+        return cleanables;
+    }
 }
