+++ core/src/test/java/org/jdbi/v3/StatementContextTest.java
@@ -22,12 +22,12 @@ import java.sql.SQLException;
 import org.jdbi.v3.tweak.ColumnMapper;
 import org.junit.Test;
 
-public class ConcreteStatementContextTest {
+public class StatementContextTest {
 
 
     @Test(expected = IllegalArgumentException.class)
     public void testShouldNotBeAbleToCombineGeneratedKeysAndConcurrentUpdatable() throws Exception {
-        final ConcreteStatementContext context = new ConcreteStatementContext();
+        final StatementContext context = new StatementContext();
 
         context.setReturningGeneratedKeys(true);
         context.setConcurrentUpdatable(true);
@@ -35,7 +35,7 @@ public class ConcreteStatementContextTest {
 
     @Test(expected = IllegalArgumentException.class)
     public void testShouldNotBeAbleToCombineConcurrentUpdatableAndGeneratedKeys() throws Exception {
-        final ConcreteStatementContext context = new ConcreteStatementContext();
+        final StatementContext context = new StatementContext();
 
         context.setConcurrentUpdatable(true);
         context.setReturningGeneratedKeys(true);
@@ -58,8 +58,7 @@ public class ConcreteStatementContextTest {
         JdbiConfig config = new JdbiConfig();
         config.mappingRegistry.addColumnMapper(mapper);
 
-        final ConcreteStatementContext context =
-                new ConcreteStatementContext(config);
+        final StatementContext context = new StatementContext(config);
 
         assertThat(context.findColumnMapperFor(Foo.class).get(), equalTo(mapper));
     }
