+++ core/src/test/java/org/jdbi/v3/TestColonStatementRewriter.java
@@ -38,7 +38,7 @@ public class TestColonStatementRewriter
     {
         return rw.rewrite(sql,
                 new Binding(),
-                new ConcreteStatementContext());
+                new StatementContext());
     }
 
     @Test
