+++ core/src/test/java/org/jdbi/v3/TestHashPrefixStatementRewriter.java
@@ -36,7 +36,7 @@ public class TestHashPrefixStatementRewriter
     public void testNewlinesOkay() throws Exception
     {
         RewrittenStatement rws = rw.rewrite("select * from something\n where id = #id", new Binding(),
-                                            new ConcreteStatementContext());
+                                            new StatementContext());
         assertEquals("select * from something\n where id = ?", rws.getSql());
     }
 
@@ -44,7 +44,7 @@ public class TestHashPrefixStatementRewriter
     public void testOddCharacters() throws Exception
     {
         RewrittenStatement rws = rw.rewrite("~* #boo '#nope' _%&^& *@ #id", new Binding(),
-                                            new ConcreteStatementContext());
+                                            new StatementContext());
         assertEquals("~* ? '#nope' _%&^& *@ ?", rws.getSql());
     }
 
@@ -52,7 +52,7 @@ public class TestHashPrefixStatementRewriter
     public void testNumbers() throws Exception
     {
         RewrittenStatement rws = rw.rewrite("#bo0 '#nope' _%&^& *@ #id", new Binding(),
-                                            new ConcreteStatementContext());
+                                            new StatementContext());
         assertEquals("? '#nope' _%&^& *@ ?", rws.getSql());
     }
 
@@ -60,7 +60,7 @@ public class TestHashPrefixStatementRewriter
     public void testDollarSignOkay() throws Exception
     {
         RewrittenStatement rws = rw.rewrite("select * from v$session", new Binding(),
-                                            new ConcreteStatementContext());
+                                            new StatementContext());
         assertEquals("select * from v$session", rws.getSql());
     }
 
@@ -68,7 +68,7 @@ public class TestHashPrefixStatementRewriter
     public void testColonIsLiteral() throws Exception
     {
         RewrittenStatement rws = rw.rewrite("select * from foo where id = :id", new Binding(),
-                                            new ConcreteStatementContext());
+                                            new StatementContext());
         assertEquals("select * from foo where id = :id", rws.getSql());
     }
 
@@ -76,7 +76,7 @@ public class TestHashPrefixStatementRewriter
     public void testBacktickOkay() throws Exception
     {
         RewrittenStatement rws = rw.rewrite("select * from `v$session", new Binding(),
-                                            new ConcreteStatementContext());
+                                            new StatementContext());
         assertEquals("select * from `v$session", rws.getSql());
     }
 
@@ -85,7 +85,7 @@ public class TestHashPrefixStatementRewriter
     {
         try {
             rw.rewrite("select * from something\n where id = #\u0087\u008e\u0092\u0097\u009c", new Binding(),
-                       new ConcreteStatementContext());
+                       new StatementContext());
 
             Assert.fail("Expected 'UnableToCreateStatementException' but got none");
         }
